pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fieldtest"
version = "0.1.0"
description = "Structured AI eval practice for any project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.20.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "click>=8.0",
]

[project.scripts]
fieldtest = "fieldtest.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
